// src/lib/api/generated/model/dayOfWeek.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Agendly.Api
 * OpenAPI spec version: 1.0
 */

export type DayOfWeek = typeof DayOfWeek[keyof typeof DayOfWeek];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DayOfWeek = {
  Sunday: 'Sunday',
  Monday: 'Monday',
  Tuesday: 'Tuesday',
  Wednesday: 'Wednesday',
  Thursday: 'Thursday',
  Friday: 'Friday',
  Saturday: 'Saturday',
} as const;